+++ internal/dataplane/translator/translator.go
@@ -56,6 +56,11 @@ type FeatureFlags struct {
 	// CombinedServicesFromDifferentHTTPRoutes indicates whether we should combine rules from different HTTPRoutes
 	// that are sharing the same combination of backends to one Kong service.
 	CombinedServicesFromDifferentHTTPRoutes bool
+	// SupportRedirectPlugin indicates whether the Kong gateway supports the `redirect` plugin.
+	// This is supported starting with Kong 3.9.
+	// If `redirect` plugin is supported, we will translate the `requestRedirect` filter to `redirect` plugin
+	// so preserving paths of request in the redirect response can be supported.
+	SupportRedirectPlugin bool
 }
 
 func NewFeatureFlags(
@@ -63,6 +68,7 @@ func NewFeatureFlags(
 	routerFlavor dpconf.RouterFlavor,
 	updateStatusFlag bool,
 	enterpriseEdition bool,
+	supportRedirectPlugin bool,
 ) FeatureFlags {
 	return FeatureFlags{
 		ReportConfiguredKubernetesObjects:       updateStatusFlag,
@@ -73,6 +79,7 @@ func NewFeatureFlags(
 		KongServiceFacade:                       featureGates.Enabled(featuregates.KongServiceFacade),
 		KongCustomEntity:                        featureGates.Enabled(featuregates.KongCustomEntity),
 		CombinedServicesFromDifferentHTTPRoutes: featureGates.Enabled(featuregates.CombinedServicesFromDifferentHTTPRoutes),
+		SupportRedirectPlugin:                   supportRedirectPlugin,
 	}
 }